add application api endpoint to start server transfer
