+++ app/Http/Controllers/Api/Application/Servers/ServerManagementController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Response;
 use App\Models\Server;
 use App\Services\Servers\SuspensionService;
 use App\Services\Servers\ReinstallServerService;
+use App\Services\Servers\TransferServerService;
 use App\Http\Requests\Api\Application\Servers\ServerWriteRequest;
 use App\Http\Controllers\Api\Application\ApplicationApiController;
 
@@ -16,7 +17,8 @@ class ServerManagementController extends ApplicationApiController
      */
     public function __construct(
         private ReinstallServerService $reinstallServerService,
-        private SuspensionService $suspensionService
+        private SuspensionService $suspensionService,
+        private TransferServerService $transferServerService,
     ) {
         parent::__construct();
     }
@@ -56,5 +58,28 @@ class ServerManagementController extends ApplicationApiController
         $this->reinstallServerService->handle($server);
 
         return $this->returnNoContent();
+    }
+
+    /**
+     * Starts a transfer of a server to a new node.
+     *
+     * @throws \App\Exceptions\DisplayException
+     * @throws \App\Exceptions\Model\DataValidationException
+     */
+    public function startTransfer(ServerWriteRequest $request, Server $server): Response
+    {
+        $validatedData = $request->validate([
+            'node_id' => 'required|exists:nodes,id',
+            'allocation_id' => 'required|bail|unique:servers|exists:allocations,id',
+            'allocation_additional' => 'nullable',
+        ]);
+
+        if ($this->transferServerService->handle($server, $validatedData)) {
+            // Transfer started
+            return new Response('', Response::HTTP_NO_CONTENT);
+        } else {
+            // Node was not viable
+            return new Response('', Response::HTTP_NOT_ACCEPTABLE);
+        }
     }
 }
